+++ PublicHeaders/SCPDiscoveryConfiguration.h
@@ -17,13 +17,10 @@
 NS_ASSUME_NONNULL_BEGIN
 
 /**
- Configuration for discovering readers.
- 
- This includes configuration of:
- - `deviceType` (`verifoneP400`, `chipper2x`, `wisePad3`)
- - `discoveryMethod` (`internet`, `bluetoothScan`, `bluetoothProximity`)
- - `locationId`
- - `simulated`
+ Configuration options for discovering readers.
+
+ Objects of this type get passed into the `Terminal.shared.discoverReaders()`
+ method to control which devices get discovered, and how.
 
  @see https://stripe.com/docs/terminal/readers/connecting
  */
@@ -34,8 +31,6 @@ NS_SWIFT_NAME(DiscoveryConfiguration)
  Initializes a discovery configuration with the given device type, selecting
  the default discovery method for the device type.
 
- Note that `simulated` is not supported by `verifoneP400`.
-
  @param deviceType          The device type to discover.
  @param simulated           Whether to use simulated discovery to discover a
  device simulator. The SDK comes with the ability to simulate behavior without
@@ -43,20 +38,17 @@ NS_SWIFT_NAME(DiscoveryConfiguration)
  end-to-end, from pairing with a reader to taking payments.
  */
 - (instancetype)initWithDeviceType:(SCPDeviceType)deviceType
-                         simulated:(BOOL)simulated;
+                         simulated:(BOOL)simulated DEPRECATED_MSG_ATTRIBUTE("All `initWithDeviceType` constructors of SCPDiscoveryConfiguration are deprecated; please use the corresponding `initWithDiscoveryMethod` constructor instead.");
 
 /**
  Initializes a discovery configuration with the given device type and discovery method.
 
- Note that not all deviceType and discovery method combinations are valid and
- not all readers support simulated mode. If the specified configuration is
- invalid, `discoverReaders` will fail.
+ Note that not all deviceType and discovery method combinations are valid.
+ If the specified configuration is invalid, `discoverReaders` will fail.
 
- | DeviceType     | DiscoveryMethod                                      | simulated         |
- | -------------- | -----------------------------------------------------|-------------------|
- | `chipper2X`    | `bluetoothProximity` (recommended), `bluetoothScan`  | `true` or `false` |
- | `verifoneP400` | `internet`                                           | `true` or `false` |
- | `wisePad3`     | `bluetoothScan`                                      | `false`           |
+ * The BBPOS Chipper 2X supports the Bluetooth Proximity and Bluetooth Scan discovery methods.
+ * The BBPOS WisePad 3 only supports the Bluetooth Scan discovery method.
+ * The Verifone P400 and BBPOS WisePOS E only support the Internet discovery method.
 
  @param deviceType          The device type to discover.
  @param discoveryMethod     The discovery method to use.
@@ -67,29 +59,58 @@ NS_SWIFT_NAME(DiscoveryConfiguration)
  */
 - (instancetype)initWithDeviceType:(SCPDeviceType)deviceType
                    discoveryMethod:(SCPDiscoveryMethod)discoveryMethod
-                         simulated:(BOOL)simulated;
+                         simulated:(BOOL)simulated DEPRECATED_MSG_ATTRIBUTE("All `initWithDeviceType` constructors of SCPDiscoveryConfiguration are deprecated; please use the corresponding `initWithDiscoveryMethod` constructor instead.");
+
+/**
+ Initializes a discovery configuration with the given discovery method. All discoverable readers
+ for that method will be discovered.
+
+ @param discoveryMethod     The discovery method to use.
+ @param simulated           Whether to use simulated discovery to discover a
+ device simulator. The SDK comes with the ability to simulate behavior without
+ using physical hardware. This makes it easy to quickly test your integration
+ end-to-end, from pairing with a reader to taking payments.
+*/
+- (instancetype)initWithDiscoveryMethod:(SCPDiscoveryMethod)discoveryMethod
+                              simulated:(BOOL)simulated;
+
+/**
+Initializes a discovery configuration with the given discovery method. All discoverable readers
+for that method will be discovered.
+
+@param discoveryMethod    The discovery method to use.
+@param locationId         The optional location ID to filter the discovered
+list to only readers at the specific location. Currently updating and creating
+locations is not supported by this SDK and will need to take place on the
+backend. Only available for the Internet discovery method.
+@param simulated           Whether to use simulated discovery to discover a
+device simulator. The SDK comes with the ability to simulate behavior without
+using physical hardware. This makes it easy to quickly test your integration
+end-to-end, from pairing with a reader to taking payments.
+*/
+- (instancetype)initWithDiscoveryMethod:(SCPDiscoveryMethod)discoveryMethod
+                             locationId:(nullable NSString *)locationId
+                              simulated:(BOOL)simulated;
 
 /**
  Initializes a discovery configuration with the given device type and
  discovery method and location ID.
 
  Not all deviceType and discovery method combinations are valid and not all
- readers support locationId or simulated mode. If the specified configuration is
- invalid, `discoverReaders` will fail. Only the following `DiscoveryConfiguration`
+ readers support locationId. If the specified configuration is invalid,
+ `discoverReaders` will fail. Only the following `DiscoveryConfiguration`
  combinations are allowed:
 
- | DeviceType     | DiscoveryMethod                                      | locationId      | simulated         |
- | -------------- | -----------------------------------------------------|-----------------|-------------------|
- | `chipper2X`    | `bluetoothProximity` (recommended), `bluetoothScan`  | `nil`           | `true` or `false` |
- | `verifoneP400` | `internet`                                           | `nil` or string | `true` or `false` |
- | `wisePad3`     | `bluetoothScan`                                      | `nil`           | `false`           |
+ * The BBPOS Chipper 2X supports the Bluetooth Proximity and Bluetooth Scan discovery methods.
+ * The BBPOS WisePad 3 only supports the Bluetooth Scan discovery method.
+ * The Verifone P400 and BBPOS WisePOS E only support the Internet discovery method.
 
  @param deviceType          The device type to discover.
  @param discoveryMethod     The discovery method to use.
  @param locationId          The optional location ID to filter the discovered
  list to only readers at the specific location. Currently updating and creating
  locations is not supported by this SDK and will need to take place on the
- backend.
+ backend. Only available for the Internet discovery method.
  @see https://stripe.com/docs/api/terminal/locations
  @param simulated           Whether to use simulated discovery to discover a
  device simulator. The SDK comes with the ability to simulate behavior without
@@ -99,7 +120,7 @@ NS_SWIFT_NAME(DiscoveryConfiguration)
 - (instancetype)initWithDeviceType:(SCPDeviceType)deviceType
                    discoveryMethod:(SCPDiscoveryMethod)discoveryMethod
                         locationId:(nullable NSString *)locationId
-                         simulated:(BOOL)simulated;
+                         simulated:(BOOL)simulated DEPRECATED_MSG_ATTRIBUTE("All `initWithDeviceType` constructors of SCPDiscoveryConfiguration are deprecated; please use the corresponding `initWithDiscoveryMethod` constructor instead.");
 
 /**
  The timeout (in seconds) after which `discoverReaders:` should fail. If the
@@ -112,9 +133,9 @@ NS_SWIFT_NAME(DiscoveryConfiguration)
 @property (nonatomic, assign, readwrite) NSUInteger timeout;
 
 /**
- The reader device type to discover.
+ The reader device type to discover. Deprecated.
  */
-@property (nonatomic, readonly) SCPDeviceType deviceType;
+@property (nonatomic, readonly) SCPDeviceType deviceType DEPRECATED_MSG_ATTRIBUTE("The `deviceType` property of SCPDiscoveryConfiguration is deprecated; please do not rely on the behavior of this property in your app.");
 
 /**
  The method by which to discover readers.