+++ CHANGELOG.md
@@ -2,8 +2,18 @@
 
 If you are using CocoaPods, update your Podfile:
 ```
-pod 'StripeTerminal', '1.3.0'
+pod 'StripeTerminal', '1.4.0'
 ```
+# 1.4.0 2020-09-03
+
+- Added [`accountType`](https://stripe.dev/stripe-terminal-ios/docs/Classes/SCPReceiptDetails.html#/c:objc(cs)SCPReceiptDetails(py)accountType) to `SCPReceiptDetails`.
+- Added [`SCPDeviceTypeWisePosE`](https://stripe.dev/stripe-terminal-ios/docs/Enums/SCPDeviceType.html#/c:@E@SCPDeviceType@SCPDeviceTypeWisePosE) to `SCPDeviceType`.
+- Deprecated the `SCPDiscoveryConfiguration` initializers that require that you specify the device type to discover and replaced them with initializers that only require that you specify the discovery method. More specifically, if you use any of these `SCPDiscoveryConfiguration` initializers, you should replace them with the corresponding versions:
+    - `initWithDeviceType:simulated:` → `initWithDiscoveryMethod:simulated:`
+    - `initWithDeviceType:discoveryMethod:simulated:` → `initWithDiscoveryMethod:simulated:`
+    - `initWithDeviceType:discoveryMethod:locationId:simulated:` → `initWithDiscoveryMethod:locationId:simulated:`
+- Added a new interface in the example app to grant easy access to receipts
+
 # 1.3.0 2020-07-28
 
 This update adds Verifone P400 support for the following methods in `StripeTerminal`:
@@ -12,8 +22,6 @@ This update adds Verifone P400 support for the following methods in `StripeTermi
 - [`readReusableCard`](https://stripe.dev/stripe-terminal-ios/docs/Classes/SCPTerminal.html#/c:objc(cs)SCPTerminal(im)readReusableCard:delegate:completion:) (only available for US merchants)
 
 This update also introduces a cosmetic change to `SCPReader.deviceSoftwareVersion`. The Key Identifier number was previously displayed as the five-character string `30000` . This and future updates will display the identifier as a six-character string: `300001`. The key itself has not changed.
-
-Not every Stripe Terminal reader supports all the functionality in the SDK. If you call a function that is not supported by the currently connected reader you will receive a `SCPErrorFeatureNotAvailableWithConnectedReader` error.
 
 # 1.2.1 2020-06-18